chore: Move helper to E2E test
